# openapi.json
{
  "components": {
    "schemas": {
...
      "ExternalFundingRequest": {
        "properties": {
          "amount_minor": {
            "exclusiveMinimum": 0.0,
            "maximum": 9000000000000.0,
            "title": "Amount Minor",
            "type": "integer"
          },
          "approved_arrangement": {
            "title": "Approved Arrangement",
            "type": "boolean"
          },
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "evidence_reference": {
            "maxLength": 500,
            "minLength": 5,
            "title": "Evidence Reference",
            "type": "string"
          }
        },
        "required": [
          "amount_minor",
          "currency",
          "evidence_reference",
          "approved_arrangement"
        ],
        "title": "ExternalFundingRequest",
        "type": "object"
      },
      "ExternalPayoutRequest": {
        "properties": {
          "approved_arrangement": {
            "const": true,
            "title": "Approved Arrangement",
            "type": "boolean"
          },
          "evidence_summary": {
            "maxLength": 2000,
            "minLength": 20,
            "title": "Evidence Summary",
            "type": "string"
          },
          "external_reference": {
            "maxLength": 255,
            "minLength": 6,
            "title": "External Reference",
            "type": "string"
          }
        },
        "required": [
          "approved_arrangement",
          "external_reference",
          "evidence_summary"
        ],
        "title": "ExternalPayoutRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PayoutAllocationCreate": {
        "properties": {
          "amount_minor": {
            "exclusiveMinimum": 0.0,
            "maximum": 9000000000000.0,
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "recipient_user_id": {
            "format": "uuid",
            "title": "Recipient User Id",
            "type": "string"
          }
        },
        "required": [
          "recipient_user_id",
          "amount_minor",
          "currency"
        ],
        "title": "PayoutAllocationCreate",
        "type": "object"
      },
      "PayoutAllocationView": {
        "properties": {
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "approved_by_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "recipient_user_id": {
            "format": "uuid",
            "title": "Recipient User Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "recipient_user_id",
          "amount_minor",
          "currency",
          "status",
          "approved_by_id",
          "created_at"
        ],
        "title": "PayoutAllocationView",
        "type": "object"
      },
      "PayoutRecordView": {
        "properties": {
          "allocation_id": {
            "format": "uuid",
            "title": "Allocation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Hash"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "provider_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Reference"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "allocation_id",
          "provider_reference",
          "status",
          "failure_reason",
          "evidence_hash",
          "created_at"
        ],
        "title": "PayoutRecordView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "PraxisAI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ops/payout-allocations/{allocation_id}/approve": {
      "post": {
        "operationId": "approve_payout_allocation_api_v1_ops_payout_allocations__allocation_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "allocation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Allocation Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "praxis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Praxis Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutAllocationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Payout Allocation",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/ops/payout-allocations/{allocation_id}/external-evidence": {
      "post": {
        "operationId": "record_payout_evidence_api_v1_ops_payout_allocations__allocation_id__external_evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "allocation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Allocation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "cookie",
            "name": "praxis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Praxis Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalPayoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutRecordView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Payout Evidence",
        "tags": [
          "billing"
        ]
      }
    },
...
    "/api/v1/ops/projects/{project_id}/external-funding": {
      "post": {
        "operationId": "record_external_funding_api_v1_ops_projects__project_id__external_funding_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "cookie",
            "name": "praxis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Praxis Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalFundingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record External Funding",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/ops/projects/{project_id}/payout-allocations": {
      "post": {
        "operationId": "create_payout_allocation_api_v1_ops_projects__project_id__payout_allocations_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "praxis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Praxis Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayoutAllocationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutAllocationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Payout Allocation",
        "tags": [
          "billing"
        ]
      }
    },
...
  }
}